sstables: move index_reader behind a virtual interface
We need this so that we can have multiple index implementations.
