+++ configure.py
@@ -853,6 +853,7 @@ scylla_core = (['message/messaging_service.cc',
                 'sstables/storage.cc',
                 'sstables/mx/partition_reversing_data_source.cc',
                 'sstables/mx/reader.cc',
+                'sstables/index_reader.cc',
                 'sstables/mx/writer.cc',
                 'sstables/kl/reader.cc',
                 'sstables/sstable_version.cc',
