+++ sstables/sstables_manager.hh
@@ -23,6 +23,7 @@
 #include "locator/host_id.hh"
 #include "reader_concurrency_semaphore.hh"
 #include "utils/s3/creds.hh"
+#include "sstables_registry.hh"
 #include <boost/intrusive/list.hpp>
 
 namespace db {
